requirements.txt
Django
djangorestframework 
grpcio
grpcio-tools
django-grpc-framework
psycopg2
protobuf
python-dotenv
google-cloud-storage
google-cloud-speech 
google-cloud-texttospeech 
google-cloud-aiplatform
pydub
vertexai

# transformers
# torch
# torchaudio
# soundfile
# datasets
# librosa
# TTS